Add files via upload
This script downloads the data from Missing Migrants Project and the Migrant files, reads the files, harmonizes them, joints them, and computes the fractionalization index HHI following https://de.wikipedia.org/wiki/Herfindahl-Index. 

It creates two files at the end: daily and monthly counts of death by routes (it also includes the western africa route).

The script also takes the data on total number of survivors, which is available from the MMP from 2014 onwards, but not before in the Migrant Files.

Finally it creates a plot of the created variables using the monthly aggregates.